_archive/Agents-Tribunal/requirements.txt
mcp[cli]
pdf2image
Pillow
nbformat
requests
PyMuPDF
google-generativeai
numpy
scipy
matplotlib


